Refactor out constant product curve
