+++ token-swap/program/src/curve/cp.rs
@@ -1,4 +1,4 @@
-//! Simple constant 1:1 swap curve
+//! The Uniswap invariant calculator.
 
 use solana_program::{
     program_error::ProgramError,
@@ -7,14 +7,14 @@ use solana_program::{
 
 use arrayref::{array_mut_ref, array_ref, array_refs, mut_array_refs};
 use std::convert::{TryFrom};
-use super::shared::{CurveCalculator, DynPack, SwapResult, calculate_fee};
+use super::shared::{CurveCalculator, DynPack, SwapResult, calculate_fee, map_zero_to_none};
 
-/// FlatCurve struct implementing CurveCalculator
+/// ConstantProductCurve struct implementing CurveCalculator
 #[derive(Clone, Debug, Default, PartialEq)]
-pub struct FlatCurve {
-    /// Fee numerator
+pub struct ConstantProductCurve {
+    /// Trade fee numerator
     pub trade_fee_numerator: u64,
-    /// Fee denominator
+    /// Trade fee denominator
     pub trade_fee_denominator: u64,
     /// Owner trade fee numerator
     pub owner_trade_fee_numerator: u64,
@@ -30,8 +30,8 @@ pub struct FlatCurve {
     pub host_fee_denominator: u64,
 }
 
-impl CurveCalculator for FlatCurve {
-    /// Flat curve swap always returns 1:1 (minus fee)
+impl CurveCalculator for ConstantProductCurve {
+    /// Constant product swap ensures x * y = constant
     fn swap(
         &self,
         source_amount: u128,
@@ -39,21 +39,21 @@ impl CurveCalculator for FlatCurve {
         swap_destination_amount: u128,
     ) -> Option<SwapResult> {
         // debit the fee to calculate the amount swapped
-        let trade_fee = calculate_fee(
-            source_amount,
-            u128::try_from(self.trade_fee_numerator).ok()?,
-            u128::try_from(self.trade_fee_denominator).ok()?,
-        )?;
+        let trade_fee = self.trading_fee(source_amount)?;
         let owner_fee = calculate_fee(
             source_amount,
             u128::try_from(self.owner_trade_fee_numerator).ok()?,
             u128::try_from(self.owner_trade_fee_denominator).ok()?,
         )?;
 
-        let amount_swapped = source_amount
+        let invariant = swap_source_amount.checked_mul(swap_destination_amount)?;
+        let new_source_amount_less_fee = swap_source_amount
+            .checked_add(source_amount)?
             .checked_sub(trade_fee)?
             .checked_sub(owner_fee)?;
-        let new_destination_amount = swap_destination_amount.checked_sub(amount_swapped)?;
+        let new_destination_amount = invariant.checked_div(new_source_amount_less_fee)?;
+        let amount_swapped =
+            map_zero_to_none(swap_destination_amount.checked_sub(new_destination_amount)?)?;
 
         // actually add the whole amount coming in
         let new_source_amount = swap_source_amount.checked_add(source_amount)?;
@@ -75,6 +75,15 @@ impl CurveCalculator for FlatCurve {
         )
     }
 
+    /// Calculate the trading fee in trading tokens
+    fn trading_fee(&self, trading_tokens: u128) -> Option<u128> {
+        calculate_fee(
+            trading_tokens,
+            u128::try_from(self.trade_fee_numerator).ok()?,
+            u128::try_from(self.trade_fee_denominator).ok()?,
+        )
+    }
+
     /// Calculate the host fee based on the owner fee, only used in production
     /// situations where a program is hosted by multiple frontends
     fn host_fee(&self, owner_fee: u128) -> Option<u128> {
@@ -87,15 +96,15 @@ impl CurveCalculator for FlatCurve {
 }
 
 /// IsInitialized is required to use `Pack::pack` and `Pack::unpack`
-impl IsInitialized for FlatCurve {
+impl IsInitialized for ConstantProductCurve {
     fn is_initialized(&self) -> bool {
         true
     }
 }
-impl Sealed for FlatCurve {}
-impl Pack for FlatCurve {
+impl Sealed for ConstantProductCurve {}
+impl Pack for ConstantProductCurve {
     const LEN: usize = 64;
-    fn unpack_from_slice(input: &[u8]) -> Result<FlatCurve, ProgramError> {
+    fn unpack_from_slice(input: &[u8]) -> Result<ConstantProductCurve, ProgramError> {
         let input = array_ref![input, 0, 64];
         #[allow(clippy::ptr_offset_with_cast)]
         let (
@@ -125,7 +134,7 @@ impl Pack for FlatCurve {
     }
 }
 
-impl DynPack for FlatCurve {
+impl DynPack for ConstantProductCurve {
     fn pack_into_slice(&self, output: &mut [u8]) {
         let output = array_mut_ref![output, 0, 64];
         let (
@@ -147,86 +156,4 @@ impl DynPack for FlatCurve {
         *host_fee_numerator = self.host_fee_numerator.to_le_bytes();
         *host_fee_denominator = self.host_fee_denominator.to_le_bytes();
     }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn flat_swap_calculation() {
-        let swap_source_amount = 1000;
-        let swap_destination_amount = 50000;
-        let trade_fee_numerator = 1;
-        let trade_fee_denominator = 100;
-        let owner_trade_fee_numerator = 2;
-        let owner_trade_fee_denominator = 100;
-        let owner_withdraw_fee_numerator = 2;
-        let owner_withdraw_fee_denominator = 100;
-        let host_fee_numerator = 2;
-        let host_fee_denominator = 100;
-        let source_amount: u128 = 100;
-        let curve = FlatCurve {
-            trade_fee_numerator,
-            trade_fee_denominator,
-            owner_trade_fee_numerator,
-            owner_trade_fee_denominator,
-            owner_withdraw_fee_numerator,
-            owner_withdraw_fee_denominator,
-            host_fee_numerator,
-            host_fee_denominator,
-        };
-        let result = curve
-            .swap(source_amount, swap_source_amount, swap_destination_amount)
-            .unwrap();
-        let amount_swapped = 97;
-        assert_eq!(result.new_source_amount, 1100);
-        assert_eq!(result.amount_swapped, amount_swapped);
-        assert_eq!(result.trade_fee, 1);
-        assert_eq!(result.owner_fee, 2);
-        assert_eq!(
-            result.new_destination_amount,
-            swap_destination_amount - amount_swapped
-        );
-    }
-
-    #[test]
-    fn pack_flat_curve() {
-        let trade_fee_numerator = 1;
-        let trade_fee_denominator = 4;
-        let owner_trade_fee_numerator = 2;
-        let owner_trade_fee_denominator = 5;
-        let owner_withdraw_fee_numerator = 4;
-        let owner_withdraw_fee_denominator = 10;
-        let host_fee_numerator = 4;
-        let host_fee_denominator = 10;
-        let curve = FlatCurve {
-            trade_fee_numerator,
-            trade_fee_denominator,
-            owner_trade_fee_numerator,
-            owner_trade_fee_denominator,
-            owner_withdraw_fee_numerator,
-            owner_withdraw_fee_denominator,
-            host_fee_numerator,
-            host_fee_denominator,
-        };
-
-        let mut packed = [0u8; FlatCurve::LEN];
-        Pack::pack_into_slice(&curve, &mut packed[..]);
-        let unpacked = FlatCurve::unpack(&packed).unwrap();
-        assert_eq!(curve, unpacked);
-
-        let mut packed = vec![];
-        packed.extend_from_slice(&trade_fee_numerator.to_le_bytes());
-        packed.extend_from_slice(&trade_fee_denominator.to_le_bytes());
-        packed.extend_from_slice(&owner_trade_fee_numerator.to_le_bytes());
-        packed.extend_from_slice(&owner_trade_fee_denominator.to_le_bytes());
-        packed.extend_from_slice(&owner_withdraw_fee_numerator.to_le_bytes());
-        packed.extend_from_slice(&owner_withdraw_fee_denominator.to_le_bytes());
-        packed.extend_from_slice(&host_fee_numerator.to_le_bytes());
-        packed.extend_from_slice(&host_fee_denominator.to_le_bytes());
-        let unpacked = FlatCurve::unpack(&packed).unwrap();
-        assert_eq!(curve, unpacked);
-    }
-
 }
