+++ token-swap/program/src/curve/mod.rs
@@ -1,4 +1,5 @@
 //! ...
 
-pub mod shared;
+pub mod cp;
 pub mod flat;
+pub mod shared;
