+++ token-swap/program/src/processor.rs
@@ -733,9 +733,10 @@ fn to_u64(val: u128) -> Result<u64, SwapError> {
 mod tests {
     use super::*;
     use crate::{
+        curve::cp::ConstantProductCurve,
         curve::flat::FlatCurve,
         curve::shared::{
-            ConstantProductCurve, CurveCalculator, CurveType, INITIAL_SWAP_POOL_AMOUNT,
+            CurveCalculator, CurveType, INITIAL_SWAP_POOL_AMOUNT,
         },
         instruction::{deposit, initialize, swap, withdraw},
     };
